Refactored away consumerIndexCache and replaced it with producerLimit.
This pattern is more akin to the producerLookAhead in the SpscArrayQueue.
It affords us some opportunity to remove some calculations from the hot
paths when batch processing elements.

Sadly the offerIfBelowThreshold is a counter intuitive method. Refactoring
it is difficult without changing the semantics. I'd suggest the semantics
of this method given the javdoc & test cases are rather confusing and
perhaps the semantics need reviewing. For backwards compatibility I've
preserved the existing semantics.

Removed redundant imports that are pulled in through inheritence.