Added profile "eclipse-dev"
This profile may be enabled in the Eclipse IDE to help the Maven
integration to decide which compiler version to use.

This is optional and non-functional. It comes with the risks that you may
not notice if the source violates Java 6 syntax until you build through
maven. On the other hand if you are doing eclipse development and not
toggling the compiler versions you probably aren't running the tests.