Resolve null dereference warnings
The warning is spurious but resolution seems reasonable

* https://lgtm.com/projects/g/JCTools/JCTools/snapshot/2c7ccd9a614c23e98493bfbd2428f9cd14baa2a4/files/jctools-experimental/src/main/java/org/jctools/queues/unpadded/BaseMpscLinkedUnpaddedArrayQueue.java?sort=name&dir=ASC&mode=heatmap#xf259fa0128f03aba:1
* https://lgtm.com/projects/g/JCTools/JCTools/snapshot/2c7ccd9a614c23e98493bfbd2428f9cd14baa2a4/files/jctools-core/src/main/java/org/jctools/queues/BaseMpscLinkedArrayQueue.java?sort=name&dir=ASC&mode=heatmap#xf259fa0128f03aba:1

The analysis claims to use the assert at the beginning of the function
to infer `s` may be null.
Upon inspection it appears to tautology, I think from usage it was
intended to assert that there must be one null, and one not null
argument for `e` and `s`.